Protobuf Schema Upate
migrate protobuf schema from custom solution to native supported version of protobuf

+++ lib/src/data_structures/utxo_pool.dart
@@ -121,7 +121,7 @@ class UtxoPool {
 
     // total the output value
     outputs.forEach((output) {
-      outputValue += output.value;
+      outputValue += output.value.toInt();
     });
     List<Utxo> selectedUtxos = [];
     print('selecting Utxos to cover ${nanoWitToWit(outputValue)}');
